Addressing PR suggestions
